refactored context menu handling into own class
